variaveiis e tipos de dados

+++ aulas/module01/media.py
@@ -2,10 +2,11 @@ nota01 = 7
 nota02 = 8
 nota03 = 9
 
-media = (nota01 + nota02 + nota03) / 3
+media = int((nota01 + nota02 + nota03) / 3)
+
 if(media >= 7):
     print("Aprovado")
-elif(media >= 5) and (media < 7):
+elif(media >= 5) and (media < 7): # (else if)
     print("Recuperação")
 else:
     print('Reprovado')
